+++ internal/worker/worker.go
@@ -2,7 +2,11 @@ package worker
 
 import (
 	"context"
-	
+	"errors"
+	"math/rand"
+
+	"distributed-job-system/internal/jobs"
+	"distributed-job-system/internal/logger"
 	"distributed-job-system/internal/queue"
 	"log"
 	"time"
@@ -11,17 +15,23 @@ import (
 type Worker struct{
 	ID int
 	Queue *queue.RedisQueue
+	MaxRetries int
 }
+func NewWorker(id int,q *queue.RedisQueue,maxRetries int,) *Worker {
 
-func NewWorker(id int, q *queue.RedisQueue) *Worker{
-	return &Worker{
-		ID: id,
-		Queue: q,
-	}
+    return &Worker{
+        ID: id,
+        Queue: q,
+        MaxRetries: maxRetries,
+    }
 }
 
 func (w *Worker) Start(ctx context.Context) {
-	log.Printf("[Worker-%d] Started", w.ID)
+
+	logger.Log.Info(
+		"Worker started",
+		"worker", w.ID,
+	)
 
 	for {
 
@@ -32,29 +42,90 @@ func (w *Worker) Start(ctx context.Context) {
 			return
 
 		default:
-			job, err := w.Queue.Pop(ctx)
+		job, err := w.Queue.Pop(ctx)
 
-			if err != nil{
-				log.Println("Pop error:", err)
-				time.Sleep(time.Second)
+		if err != nil {
+			logger.Log.Error(
+				"Failed to pop job",
+				"worker",w.ID,
+				"error",err,
+			)
+			continue
+		}
+
+		Recievedjob := jobs.Job{
+			ID: job.ID,
+			Type: job.Type,
+			Payload: job.Payload,
+			RetryCount: job.RetryCount,
+			MaxRetries: w.MaxRetries,
+			CreatedAt: time.Now(),
+		}
+
+		err = Process(Recievedjob)
+
+		if err != nil {
+
+			job.RetryCount++
+
+			if job.RetryCount <= w.MaxRetries {
+
+				logger.Log.Info(
+					"Retrying job",
+					"worker", w.ID,
+					"job", job.ID,
+					"Total retries", job.RetryCount,
+				)
+
+				Updatedjob := jobs.Job{
+					ID: job.ID,
+					Type: job.Type,
+					Payload: job.Payload,
+					RetryCount: job.RetryCount,
+					MaxRetries: w.MaxRetries,
+					CreatedAt: time.Now(),
+				}
+
+				err := w.Queue.Push(ctx,Updatedjob)
+
+				if err != nil {
+					logger.Log.Error(
+						"Failed to Requeue job",
+						"worker", w.ID,
+						"job_id", job.ID,
+						"Error", err,
+					)
+					continue
+				}
+
+					continue
+				}
+
+				logger.Log.Error(
+					"job Permanently failed",
+					"Worker", w.ID,
+					"job", job.ID,
+				)
+
 				continue
 			}
 
-			log.Printf(
-					"[Worker-%d] Processing Job %s (%s)",
-					w.ID,
-					job.ID,
-					job.Type,
-				)
-
-			time.Sleep(2 *time.Second)
-
-			log.Printf(
-				"[Worker-%d] Finished Job %s",
-				w.ID,
-				job.ID,
+			logger.Log.Info(
+				"job processed succesful",
+				"worker", w.ID,
+				"job", job.ID,
 			)
-
 		}
 	}
+}
+
+func Process(job jobs.Job) error{
+
+	time.Sleep(2 * time.Second)
+
+	if rand.Intn(10) < 3{
+		return errors.New("simulated processing failure")
+	}
+
+	return nil
 }